Initial_Test/requirement.txt
langchain_huggingface
python-dotenv